+++ src/Utility/BackupExport.php
@@ -128,7 +128,7 @@ class BackupExport
      * @uses $connection
      * @return string File path
      */
-    protected function _storeAuth()
+    private function _storeAuth()
     {
         $auth = tempnam(sys_get_temp_dir(), 'auth');